chore: usage of hooks
