+++ src/hooks/index.ts
@@ -1,3 +1,5 @@
 export * from './context';
 export * from './useAsyncOperation';
 export * from './useVoting';
+export * from './useWethBalance';
+export * from './useBridgeOut';
